requirements.txt
fastapi
uvicorn
python-multipart
loguru
pdfplumber
python-docx
chromadb
sentence-transformers
pytest
llama-cpp-python
huggingface_hub
openai-whisper
kokoro
soundfile
